rename skills to abilities in internal data structure
