+++ CHANGELOG.md
@@ -8,3 +8,4 @@
 - Renamed _baseattributes_ to _attributes_ on actor data (#27)
 - Added dice icons to rollable items and optimized css in chatcard presentation 
 - Removed Homebrew Inspiration points, as this should be a module (#40)
+- Reanmed _skill_ to _ability_ in internal data structure
