+++ module/data/item/social.mjs
@@ -1,4 +1,4 @@
-import HowToBeAHeroSkill from "./skill.mjs";
+import HowToBeAHeroSkill from "./ability.mjs";
 
 export default class HowToBeAHeroSocial extends HowToBeAHeroSkill {
