+++ module/helpers/settings.mjs
@@ -210,9 +210,9 @@ export function registerSystemSettings() {
       keepPhysical: false,
       keepMental: false,
       keepSaves: false,
-      keepSkills: false,
+      keepAbilities: false,
       mergeSaves: false,
-      mergeSkills: false,
+      mergeAbilities: false,
       keepClass: false,
       keepFeats: false,
       keepSpells: false,
